add DB scheme Завершено ревью пром задания ФЗ 12 добавить БД H2 добавить схему для создания таблиц добавить схему для создания словарей - рейтинг, жанр, статус дружбы add set genres to class Film add jdbc to pom add UserRepository add UserService add UserController finish UserController and all logic - postman passed

+++ src/main/resources/schema.sql
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS films (
                                      release_date DATE NOT NULL CHECK (films.release_date >= '1895-12-28'),
                                      duration INTEGER NOT NULL CHECK (duration > 0),
                                      description VARCHAR(200),
-                                     mpa INT NOT NULL references mpa(id) ON DELETE RESTRICT
+                                     mpa BIGINT NOT NULL references mpa(id) ON DELETE RESTRICT
 );
 
 CREATE TABLE IF NOT EXISTS films_likes (
